+++ unlock-js/src/v01/updateKeyPrice.js
@@ -7,8 +7,9 @@ import TransactionTypes from '../transactionTypes'
  * @param {object} params
  * - {PropTypes.address} lockAddress : address of the lock for which we update the price
  * - {string} keyPrice : new price for the lock, as a string (as wei)
+ * @param {function} callback invoked with the transaction hash
  */
-export default async function({ lockAddress, keyPrice }) {
+export default async function({ lockAddress, keyPrice }, callback) {
   const lockContract = await this.getLockContract(lockAddress)
   const transactionPromise = lockContract['updateKeyPrice(uint256)'](
     utils.toWei(keyPrice, 'ether'),
@@ -20,6 +21,10 @@ export default async function({ lockAddress, keyPrice }) {
     transactionPromise,
     TransactionTypes.UPDATE_KEY_PRICE
   )
+
+  if (callback) {
+    callback(null, hash)
+  }
 
   // Let's now wait for the keyPrice to have been changed before we return it
   const receipt = await this.provider.waitForTransaction(hash)