+++ unlock-js/src/v10/purchaseKey.js
@@ -1,7 +1,7 @@
 import utils from '../utils'
 import { GAS_AMOUNTS, ZERO } from '../constants'
 import TransactionTypes from '../transactionTypes'
-import { approveTransfer, getErc20Decimals } from '../erc20'
+import { approveTransfer, getErc20Decimals, getAllowance } from '../erc20'
 
 /**
  * Purchase key function. This implementation requires the following
@@ -11,15 +11,12 @@ import { approveTransfer, getErc20Decimals } from '../erc20'
  * - {string} keyPrice
  * - {PropTypes.address} erc20Address
  * - {number} decimals
- * @return {string} hash of the transaction
+ * @param {function} callback invoked with the transaction hash
  */
-export default async function({
-  lockAddress,
-  owner,
-  keyPrice,
-  erc20Address,
-  decimals,
-}) {
+export default async function(
+  { lockAddress, owner, keyPrice, erc20Address, decimals },
+  callback
+) {
   const lockContract = await this.getLockContract(lockAddress)
 
   if (!erc20Address || erc20Address !== ZERO) {
@@ -43,12 +40,19 @@ export default async function({
   }
 
   if (erc20Address && erc20Address !== ZERO) {
-    await approveTransfer(
+    const approvedAmount = await getAllowance(
       erc20Address,
       lockAddress,
-      actualAmount,
       this.provider
     )
+    if (approvedAmount < actualAmount) {
+      await approveTransfer(
+        erc20Address,
+        lockAddress,
+        actualAmount,
+        this.provider
+      )
+    }
   } else {
     purchaseForOptions.value = actualAmount
   }
@@ -65,6 +69,9 @@ export default async function({
   // Let's now wait for the transaction to go thru to return the token id
   const receipt = await this.provider.waitForTransaction(hash)
   const parser = lockContract.interface
+  if (callback) {
+    callback(null, hash)
+  }
 
   const transferEvent = receipt.logs
     .map(log => {